% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_shrink.R
\name{mat_shrink}
\alias{mat_shrink}
\title{Soft threshholding operator.}
\usage{
mat_shrink(K, tau)
}
\arguments{
\item{K}{matrix}

\item{tau}{regularization parameter}
}
\value{
Matrix
}
\description{
Applies the shrinkage operator for singular value tresholding.
}
\examples{
mat_shrink(matrix(c(1,0,0,0,1,1,1,1,1), nrow=3, ncol=3, byrow=TRUE),0.35)
}
